Pushed up methods from TolerantReflectionMethod
