+++ lib/Bridge/TolerantParser/Reflection/ReflectionMethod.php
@@ -78,11 +78,6 @@ class ReflectionMethod extends AbstractReflectionClassMember implements CoreRefl
         return $this->node->getName();
     }
 
-    public function frame(): Frame
-    {
-        return $this->serviceLocator->frameBuilder()->build($this->node);
-    }
-
     public function declaringClass(): ReflectionClassLike
     {
         $classDeclaration = $this->node->getFirstAncestor(ClassLike::class);
@@ -101,45 +96,9 @@ class ReflectionMethod extends AbstractReflectionClassMember implements CoreRefl
         return $this->serviceLocator->reflector()->reflectClassLike(ClassName::fromString($class));
     }
 
-    public function isAbstract(): bool
-    {
-        foreach ($this->node->modifiers as $token) {
-            if ($token->kind === TokenKind::AbstractKeyword) {
-                return true;
-            }
-        }
-
-        return false;
-    }
-
-    public function isStatic(): bool
-    {
-        return $this->node->isStatic();
-    }
-
     public function parameters(): CoreReflectionParameterCollection
     {
         return ReflectionParameterCollection::fromMethodDeclaration($this->serviceLocator, $this->node, $this);
-    }
-
-    public function docblock(): DocBlock
-    {
-        return $this->serviceLocator->docblockFactory()->create($this->node->getLeadingCommentAndWhitespaceText());
-    }
-
-    public function visibility(): Visibility
-    {
-        foreach ($this->node->modifiers as $token) {
-            if ($token->kind === TokenKind::PrivateKeyword) {
-                return Visibility::private();
-            }
-
-            if ($token->kind === TokenKind::ProtectedKeyword) {
-                return Visibility::protected();
-            }
-        }
-
-        return Visibility::public();
     }
 
     public function inferredReturnTypes(): Types
@@ -148,6 +107,11 @@ class ReflectionMethod extends AbstractReflectionClassMember implements CoreRefl
     }
 
     public function returnType(): Type
+    {
+        return $this->type();
+    }
+
+    public function type(): Type
     {
         return $this->memberTypeResolver->resolve($this->class()->name(), $this->node, $this->node->returnType);
     }
