+++ lib/Core/Reflection/ReflectionConstant.php
@@ -5,15 +5,6 @@ namespace Phpactor\WorseReflection\Core\Reflection;
 use Phpactor\WorseReflection\Core\Position;
 use Phpactor\WorseReflection\Core\Type;
 
-interface ReflectionConstant
+interface ReflectionConstant extends ReflectionMember
 {
-    public function position(): Position;
-
-    public function declaringClass(): ReflectionClassLike;
-
-    public function class(): ReflectionClassLike;
-
-    public function name();
-
-    public function type(): Type;
 }
